Angular: Isolate angular modal support from react modal support
