+++ public/app/core/services/ModalManager.ts
@@ -1,36 +1,20 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import coreModule from 'app/angular/core_module';
 import appEvents from 'app/core/app_events';
-import { GrafanaRootScope } from 'app/angular/GrafanaCtrl';
 import { AngularModalProxy } from '../components/modals/AngularModalProxy';
 import { provideTheme } from '../utils/ConfigProvider';
-import {
-  HideModalEvent,
-  ShowConfirmModalEvent,
-  ShowConfirmModalPayload,
-  ShowModalEvent,
-  ShowModalReactEvent,
-} from '../../types/events';
+import { ShowConfirmModalEvent, ShowConfirmModalPayload, ShowModalReactEvent } from '../../types/events';
 import { ConfirmModal, ConfirmModalProps } from '@grafana/ui';
-import { deprecationWarning, textUtil } from '@grafana/data';
+import { textUtil } from '@grafana/data';
 import { CopyPanelEvent } from '@grafana/runtime';
 import { copyPanel } from 'app/features/dashboard/utils/panel';
 
-export class UtilSrv {
-  modalScope: any;
+export class ModalManager {
   reactModalRoot = document.body;
   reactModalNode = document.createElement('div');
 
-  /** @ngInject */
-  constructor(private $rootScope: GrafanaRootScope, private $modal: any) {
-    this.reactModalNode.setAttribute('id', 'angular2ReactModalRoot');
-  }
-
   init() {
-    appEvents.subscribe(ShowModalEvent, (e) => this.showModal(e.payload));
-    appEvents.subscribe(HideModalEvent, this.hideModal.bind(this));
     appEvents.subscribe(ShowConfirmModalEvent, (e) => this.showConfirmModal(e.payload));
     appEvents.subscribe(ShowModalReactEvent, (e) => this.showModalReact(e.payload));
     appEvents.subscribe(CopyPanelEvent, (e) => copyPanel(e.payload));
@@ -56,50 +40,6 @@ export class UtilSrv {
     ReactDOM.unmountComponentAtNode(this.reactModalNode);
     this.reactModalRoot.removeChild(this.reactModalNode);
   };
-
-  /**
-   * @deprecated use showModalReact instead that has this capability built in
-   */
-  hideModal() {
-    deprecationWarning('UtilSrv', 'hideModal', 'showModalReact');
-    if (this.modalScope && this.modalScope.dismiss) {
-      this.modalScope.dismiss();
-    }
-  }
-
-  /**
-   * @deprecated use showModalReact instead
-   */
-  showModal(options: any) {
-    deprecationWarning('UtilSrv', 'showModal', 'showModalReact');
-    if (this.modalScope && this.modalScope.dismiss) {
-      this.modalScope.dismiss();
-    }
-
-    this.modalScope = options.scope;
-
-    if (options.model) {
-      this.modalScope = this.$rootScope.$new();
-      this.modalScope.model = options.model;
-    } else if (!this.modalScope) {
-      this.modalScope = this.$rootScope.$new();
-    }
-
-    const modal = this.$modal({
-      modalClass: options.modalClass,
-      template: options.src,
-      templateHtml: options.templateHtml,
-      persist: false,
-      show: false,
-      scope: this.modalScope,
-      keyboard: false,
-      backdrop: options.backdrop,
-    });
-
-    Promise.resolve(modal).then((modalEl) => {
-      modalEl.modal('show');
-    });
-  }
 
   showConfirmModal(payload: ShowConfirmModalPayload) {
     const {
@@ -147,5 +87,3 @@ export class UtilSrv {
     ReactDOM.render(elem, this.reactModalNode);
   }
 }
-
-coreModule.service('utilSrv', UtilSrv);
