--- public/app/core/services/all.ts
@@ -1,3 +1,0 @@
-import './util_srv';
-import './context_srv';
-import './backend_srv';
